+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 0.6.1 / TBD
 ==================
+Expose Elasticsearch timeout as environment variable
 
 0.6.0 / 2017-12-14
 ==================